+++ src/generators/model/Generator.php
@@ -105,7 +105,9 @@ class Generator extends \yii\gii\generators\model\Generator
      * @var string form field for selecting and loading saved gii forms 
      */
     public $savedForm;
-    
+
+    public $messageCategory = 'models';
+
     protected $classNames2;
 
     /**
@@ -160,6 +162,7 @@ class Generator extends \yii\gii\generators\model\Generator
             'baseClass',
             'db',
             'generateRelations',
+            'generateRelationsFromCurrentSchema',
             'generateLabelsFromComments',
             'generateHintsFromComments',
             'generateModelClass',