+++ src/generators/model/default/model-extended.php
@@ -19,12 +19,33 @@ namespace <?= $generator->ns ?>;
 
 use Yii;
 use \<?= $generator->ns ?>\base\<?= $className ?> as Base<?= $className ?>;
+use yii\helpers\ArrayHelper;
 
 /**
  * This is the model class for table "<?= $tableName ?>".
  */
 class <?= $className ?> extends Base<?= $className . "\n" ?>
 {
+
+public function behaviors()
+    {
+        return ArrayHelper::merge(
+            parent::behaviors(),
+            [
+                # custom behaviors
+            ]
+        );
+    }
+
+    public function rules()
+    {
+        return ArrayHelper::merge(
+             parent::rules(),
+             [
+                  # custom validation rules
+             ]
+        );
+    }
 
     public static $listBoxData;
 
@@ -83,4 +104,5 @@ class <?= $className ?> extends Base<?= $className . "\n" ?>
         return self::$listBoxData = \yii\helpers\ArrayHelper::map($rows,'id', 'name');        
     }        
 <?php endif; ?>
+
 }